refactor(Http11Processor): HttpResponse를 사용하도록 변경
