+++ tomcat/src/main/java/org/apache/coyote/http11/HttpRequest.java
@@ -34,4 +34,8 @@ public class HttpRequest {
     public Map<String, String> getQueryParamMap() {
         return requestLine.getQueryParamMap();
     }
+
+    public HttpProtocol getProtocol() {
+        return requestLine.getProtocol();
+    }
 }
