+++ tomcat/src/main/java/org/apache/coyote/http11/RequestLine.java
@@ -44,4 +44,8 @@ public class RequestLine {
     public Map<String, String> getQueryParamMap() {
         return requestTarget.getQueryParamMap();
     }
+
+    public HttpProtocol getProtocol() {
+        return httpProtocol;
+    }
 }
